Add Notifications Provider.
